backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pg8000==1.31.2
redis==6.4.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
APScheduler==3.11.0
httpx==0.28.1
PyJWT==2.10.1
